move interceptor skip to config lib
