+++ projects/valtimo/case/src/lib/services/case-widgets-api.service.ts
@@ -15,10 +15,9 @@
  */
 import {HttpClient, HttpHeaders} from '@angular/common/http';
 import {Injectable} from '@angular/core';
-import {BaseApiService, ConfigService} from '@valtimo/config';
+import {BaseApiService, ConfigService, InterceptorSkip} from '@valtimo/config';
 import {Observable} from 'rxjs';
 import {CaseWidgetsRes} from '../models';
-import {InterceptorSkip} from '@valtimo/security';
 
 @Injectable({
   providedIn: 'root',
