+++ projects/valtimo/form/src/lib/form.service.ts
@@ -18,8 +18,7 @@ import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
 import {Observable} from 'rxjs';
 import {FormioForm} from '@formio/angular';
-import {ConfigService} from '@valtimo/config';
-import {InterceptorSkip} from '@valtimo/security';
+import {ConfigService, InterceptorSkip} from '@valtimo/config';
 import {FormDefinitionOption} from './models';
 
 @Injectable({
