+++ projects/valtimo/security/src/lib/error/error.ts
@@ -14,11 +14,6 @@
  * limitations under the License.
  */
 
-import {HttpHeaders} from '@angular/common/http';
-
-export const InterceptorSkip = 'X-Skip-Interceptor';
-export const InterceptorSkipHeader = new HttpHeaders().set(InterceptorSkip, 'all');
-
 export interface Error {
   title: string;
   error: number;
